move sf-icon-base out from sf-icons
