+++ projects/ng-storefront-ui/src/lib/sf-icon-base/sf-icon-base.component.spec.ts
