+++ projects/ng-storefront-ui/src/lib/sf-icons/sf-icon-star-filled/sf-icon-star-filled.component.ts
@@ -1,6 +1,6 @@
 import { ChangeDetectionStrategy, Component, Input } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { SfIconBaseComponent } from '../sf-icon-base/sf-icon-base.component';
+import { SfIconBaseComponent } from '../../sf-icon-base/sf-icon-base.component';
 import { SfIconSize } from '../../../types/sf-icon-base.type';
 
 @Component({
